+++ src/main/java/com/yanny/ytech/configuration/recipe/HammeringRecipe.java
@@ -27,7 +27,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Consumer;
 
-public record HammeringRecipe(ResourceLocation id, Ingredient ingredient, Ingredient tool, ItemStack result) implements Recipe<Container> {
+public record HammeringRecipe(ResourceLocation id, Ingredient ingredient, Ingredient tool, int hitCount, ItemStack result) implements Recipe<Container> {
     @Override
     public boolean matches(@NotNull Container container, @NotNull Level level) {
         return ingredient.test(container.getItem(0));
@@ -74,32 +74,36 @@ public record HammeringRecipe(ResourceLocation id, Ingredient ingredient, Ingred
         public HammeringRecipe fromJson(@NotNull ResourceLocation recipeId, @NotNull JsonObject serializedRecipe) {
             Ingredient ingredient = Ingredient.fromJson(serializedRecipe.get("ingredient"), false);
             Ingredient tool = Ingredient.fromJson(serializedRecipe.get("tool"), true);
+            int hitCount = GsonHelper.getAsInt(serializedRecipe, "hitCount");
             ItemStack result = ShapedRecipe.itemStackFromJson(GsonHelper.getAsJsonObject(serializedRecipe, "result"));
-            return new HammeringRecipe(recipeId, ingredient, tool, result);
+            return new HammeringRecipe(recipeId, ingredient, tool, hitCount, result);
         }
 
         @Override
         public @Nullable HammeringRecipe fromNetwork(@NotNull ResourceLocation recipeId, @NotNull FriendlyByteBuf buffer) {
             Ingredient ingredient = Ingredient.fromNetwork(buffer);
             Ingredient tool = Ingredient.fromNetwork(buffer);
+            int hitCount = buffer.readInt();
             ItemStack result = buffer.readItem();
-            return new HammeringRecipe(recipeId, ingredient, tool, result);
+            return new HammeringRecipe(recipeId, ingredient, tool, hitCount, result);
         }
 
         @Override
         public void toNetwork(@NotNull FriendlyByteBuf buffer, @NotNull HammeringRecipe recipe) {
             recipe.ingredient.toNetwork(buffer);
             recipe.tool.toNetwork(buffer);
+            buffer.writeInt(recipe.hitCount);
             buffer.writeItem(recipe.result);
         }
     }
 
-    public record Result(@NotNull ResourceLocation id, @NotNull Ingredient ingredient, @NotNull Ingredient tool, @NotNull Item result,
+    public record Result(@NotNull ResourceLocation id, @NotNull Ingredient ingredient, @NotNull Ingredient tool, int hitCount, @NotNull Item result,
                          @NotNull Advancement.Builder advancement, @NotNull ResourceLocation advancementId) implements FinishedRecipe {
         @Override
         public void serializeRecipeData(@NotNull JsonObject json) {
             json.add("ingredient", ingredient.toJson());
             json.add("tool", tool.toJson());
+            json.addProperty("hitCount", hitCount);
 
             JsonObject resultItemStack = new JsonObject();
             resultItemStack.addProperty("item", Utils.loc(result).toString());
@@ -134,20 +138,22 @@ public record HammeringRecipe(ResourceLocation id, Ingredient ingredient, Ingred
     public static class Builder implements RecipeBuilder {
         private final Ingredient ingredient;
         private Ingredient tool = Ingredient.EMPTY;
+        int hitCount;
         private final Item result;
         private final Advancement.Builder advancement = Advancement.Builder.recipeAdvancement();
 
-        Builder(@NotNull Ingredient ingredient, @NotNull Item result) {
+        Builder(@NotNull Ingredient ingredient, int hitCount, @NotNull Item result) {
             this.ingredient = ingredient;
+            this.hitCount = hitCount;
             this.result = result;
         }
 
-        public static Builder hammering(@NotNull TagKey<Item> input, @NotNull Item result) {
-            return new Builder(Ingredient.of(input), result);
+        public static Builder hammering(@NotNull TagKey<Item> input, int hitCount, @NotNull Item result) {
+            return new Builder(Ingredient.of(input), hitCount, result);
         }
 
-        public static Builder hammering(@NotNull ItemLike input, @NotNull Item result) {
-            return new Builder(Ingredient.of(input), result);
+        public static Builder hammering(@NotNull ItemLike input, int hitCount, @NotNull Item result) {
+            return new Builder(Ingredient.of(input), hitCount, result);
         }
 
         public Builder tool(@NotNull Ingredient tool) {
@@ -179,7 +185,7 @@ public record HammeringRecipe(ResourceLocation id, Ingredient ingredient, Ingred
             ensureValid(recipeId);
             advancement.parent(ROOT_RECIPE_ADVANCEMENT).addCriterion("has_the_recipe",
                     RecipeUnlockedTrigger.unlocked(recipeId)).rewards(AdvancementRewards.Builder.recipe(recipeId)).requirements(RequirementsStrategy.OR);
-            finishedRecipeConsumer.accept(new HammeringRecipe.Result(recipeId, ingredient, tool, result, advancement, recipeId.withPrefix("recipes/hammering/")));
+            finishedRecipeConsumer.accept(new HammeringRecipe.Result(recipeId, ingredient, tool, hitCount, result, advancement, recipeId.withPrefix("recipes/hammering/")));
         }
 
         //Makes sure that this recipe is valid and obtainable.